DFE-1021 Adding swagger docs - note that we have to add a baseurl as it currently stands. I think there is scope for moving to the top level / instead of /api and /tools as this should make things simpler
