+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/FileController.cs
@@ -7,6 +7,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Controllers
 {
+    [ApiExplorerSettings(IgnoreApi=true)]
     [Authorize]
     public class FileController : Controller
     {
