+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/PublicationsController.cs
@@ -11,6 +11,7 @@ using Microsoft.AspNetCore.Authorization;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Controllers
 {
+    [ApiExplorerSettings(IgnoreApi=true)]
     [Authorize]
     public class PublicationsController : Controller
     {
