+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/PublishingController.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Controllers
 {
+    [ApiExplorerSettings(IgnoreApi=true)]
     [Authorize]
     public class PublishingController : Controller
     {
