+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Controllers/TopicsController.cs
@@ -10,6 +10,7 @@ using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
 
 namespace GovUk.Education.ExploreEducationStatistics.Admin.Controllers
 {
+    [ApiExplorerSettings(IgnoreApi=true)]
     public class TopicsController : Controller
     {
         private readonly ApplicationDbContext _context;
